updated profile page and added comments
