+++ src/app/dashboard/page.tsx
@@ -1,12 +1,15 @@
-import { auth } from "@clerk/nextjs/server";
-// import { redirect } from "next/navigation";
 import HomepageScreen from "@/components/dashboard/HomepageScreen";
+import { auth } from "@clerk/nextjs/server";
+import { redirect } from "next/navigation";
 type Props = {};
-// GlacialIndifference: Use this font for titles, headings, and other high-visibility elements. It can provide strong visual impact and distinctiveness.
-// Avenir: Use this font for body text, small content, and secondary elements. It’s typically more neutral and optimized for readability, making it suitable for longer text passages.
 
+// Main dashboard home page
 const Dashboard = (props: Props) => {
     const { userId } = auth();
+
+    // If user is not found, show error page
+    if (!userId) return redirect("/login");
+
     return (
         <div>
             <HomepageScreen />
