+++ src/components/channelContents/AllGroups.tsx
@@ -184,7 +184,7 @@ const AllGroups: React.FC<AllGroupsProps> = ({ channel, currentUser }) => {
                                         height={20}
                                         className="w-10 h-10 rounded-full"
                                     />
-                                    <span className="font-semibold font-medium">
+                                    <span className="font-medium">
                                         {channel.channel_name}
                                     </span>
                                 </Link>
