--- src/components/dashboard/Sidebar/menuLink/menuLink.module.css
@@ -1,15 +1,0 @@
-.container {
-    padding: 20px;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    margin: 5px 0px;
-    border-radius: 10px;
-    color: #224c6b
-  }
-  
-  .container:hover,
-  .active {
-    background-color: white;
-    color:#224c6b
-  }
