--- src/components/dashboard/Sidebar/sidebar.module.css
@@ -1,57 +1,0 @@
-.container{
-  position: sticky;
-  top: 40px;
-}
-
-.user{
-  display: flex;
-  align-items: center;
-  gap: 20px;
-  margin-bottom: 20px;
-}
-
-.userImage{
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-.userDetail{
-  display: flex;
-  flex-direction: column;
-}
-
-.username{
-  font-weight: 500;
-}
-
-.userTitle{
-  font-size: 12px;
-  color: var(--textSoft);
-}
-
-.list{
-  list-style: none;
-}
-
-.cat{
-  font-weight: bold;
-  font-size: 13px;
-  margin: 10px 0px;
-}
-
-.logout{
-  padding: 20px;
-  margin: 5px 0px;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  cursor: pointer;
-  border-radius: 10px;
-  background: none;
-  border: none;
-  width: 100%;
-}
-
-.logout:hover{
-  background-color: white;
-}
