--- src/components/dashboard/dashboard.module.css
@@ -1,44 +1,0 @@
-.container {
-  display: flex;
-}
-
-.menu {
-  flex: 1;
-  background-color: #9ad0e3;
-  padding: 20px;
-  min-height: 100vh;
-}
-
-.content {
-  flex: 6;
-  padding: 20px;
-}
-
-.wrapper {
-  display: flex;
-  gap: 20px;
-  margin-top: 20px;
-}
-
-.main {
-  flex: 3;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.cards {
-  display: flex;
-  gap: 20px;
-  justify-content: space-between;
-}
-
-.side {
-  flex: 1;
-}
-/* Menu hidden on smaller screens */
-@media (max-width: 768px) {
-  .menu {
-      display: none;
-  }
-}
