--- src/components/dashboard/profile/singleUser/singleUser.module.css
@@ -1,60 +1,0 @@
-.container {
-    display: flex;
-    gap: 50px;
-    margin-top: 20px;
-  }
-  
-  .infoContainer {
-    flex: 1;
-    background-color: var(--bgSoft);
-    padding: 20px;
-    border-radius: 10px;
-    font-weight: bold;
-    color: var(--textSoft);
-    height: max-content;
-  }
-  
-  .imgContainer{
-    width: 100%;
-    height: 300px;
-    position: relative;
-    border-radius: 10px;
-    overflow: hidden;
-    margin-bottom: 20px;
-  }
-  
-  .formContainer {
-    flex: 3;
-    background-color: var(--bgSoft);
-    padding: 20px;
-    border-radius: 10px;
-  }
-  
-  .form{
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .form input, .form textarea, .form select{
-    padding: 20px;
-    border: 2px solid #2e374a;
-    border-radius: 5px;
-    background-color: var(--bg);
-    color: var(--text);
-    margin: 10px 0px;
-  }
-  
-  .form label{
-    font-size: 12px;
-  }
-  
-  .form button{
-    width: 100%;
-    padding: 20px;
-    background-color: teal;
-    color: var(--text);
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    margin-top: 20px;
-  }
