+++ src/lib/actions.ts
@@ -5,6 +5,44 @@ import { revalidatePath } from "next/cache";
 import { z } from "zod";
 import prisma from "./client";
 
+// Following are Channels page actions, please not modify them
+type User = {
+  id: string;
+  username: string;
+  profile_image: string | null;
+  first_name: string | null;
+  last_name: string | null;
+  organization: string | null;
+  title: string | null;
+  phone: string | null; 
+  description: string | null;
+  password: string | null;
+  personal_email: string | null;
+  graduation_year: string | null;
+  work_email: string | null;
+  createdAt: Date;
+};
+
+type Comment = {
+  id: number;
+  desc: string;
+  userId: string;
+  postId: number;
+  user: User;
+};
+
+type Post = {
+  id: number;
+  desc: string;
+  img: string | null;
+  video: string | null;
+  createdAt: Date;
+  updatedAt: Date;
+  userId: string;
+  channelId: number;
+  user: User;
+  comments: Comment[];
+};
 
 export const updateProfile = async(formData: FormData, cover: string, profile: string) => {
   const fields = Object.fromEntries(formData);
@@ -93,56 +131,6 @@ export const updateProfile = async(formData: FormData, cover: string, profile: s
     return { status: 'error', message: 'Internal Server Error' };
   }
 }
-
-// Following are Channels page actions, please not modify them
-type User = {
-  id: string;
-  username: string;
-  profile_image: string | null;
-  first_name: string | null;
-  last_name: string | null;
-  organization: string | null;
-  title: string | null;
-  phone: string | null; 
-  description: string | null;
-  password: string | null;
-  personal_email: string | null;
-  graduation_year: string | null;
-  work_email: string | null;
-  createdAt: Date;
-};
-
-type Comment = {
-  id: number;
-  desc: string;
-  userId: string;
-  postId: number;
-  user: User;
-};
-
-type Post = {
-  id: number;
-  desc: string;
-  img: string | null;
-  video: string | null;
-  createdAt: Date;
-  updatedAt: Date;
-  userId: string;
-  channelId: number;
-  user: User;
-  comments: Comment[];
-};
-
-type Channel = {
-  id: number;
-  channel_name: string;
-  channel_image: string | null;
-  channel_description: string | null;
-  users: {
-    user: User;
-  }[];
-  posts: Post[];
-};
 
 export const fetchUserById = async (userId: string):  Promise<User | null>=> {
   try {
