+++ src/types/index.tsx
@@ -1,19 +1,20 @@
 export interface User {
     id: string;
     username: string;
-    profile_image: string | null; // Updated to allow null
-    cover_image: string | null; // Updated to allow null
-    first_name: string | null; // Updated to allow null
-    last_name: string | null; // Updated to allow null
-    organization: string | null; // Updated to allow null
-    title: string | null; // Updated to allow null
-    phone: string | null; // Updated to allow null
-    description: string | null; // Updated to allow null
-    password: string | null; // Updated to allow null
-    personal_email: string | null; // Updated to allow null
-    graduation_year: string | null; // Updated to allow null
-    work_email: string | null; // Updated to allow null
-    admin: number | null;
+    profile_image: string | null;
+    cover_image: string | null;
+    first_name: string | null;
+    last_name: string | null;
+    organization: string | null;
+    title: string | null;
+    phone: string | null;
+    description: string | null;
+    password: string | null;
+    personal_email: string | null;
+    graduation_year: string | null;
+    work_email: string | null;
+    createdAt: Date;
+    admin: boolean | null;
 }
 
 export interface FileItem {
@@ -29,7 +30,7 @@ export interface FileItem {
     key?: string;
     url?: string;
 }
-  
+
 export interface FolderItem {
     id: number;
     courseId: number;
@@ -37,3 +38,36 @@ export interface FolderItem {
     createdAt: Date;
     files: FileItem[];
 }
+
+// Channel Types
+export type Comment = {
+    id: number;
+    desc: string;
+    userId: string;
+    postId: number;
+    user: User;
+};
+
+export type Post = {
+    id: number;
+    desc: string;
+    img: string | null;
+    video: string | null;
+    createdAt: Date;
+    updatedAt: Date;
+    userId: string;
+    channelId: number;
+    user: User;
+    comments: Comment[];
+};
+
+export type Channel = {
+    id: number;
+    channel_name: string;
+    channel_image: string | null;
+    channel_description: string | null;
+    users: {
+        user: User;
+    }[];
+    posts: Post[];
+};
